fix float prix produit

+++ site/src/Entity/Product.php
@@ -27,7 +27,7 @@ class Product
 
     /**
      * Prix du produit
-     * @ORM\Column(type="decimal", precision = 2)
+     * @ORM\Column(type="decimal", precision=5, scale=2)
      *
      * @Assert\Positive
      */
@@ -59,12 +59,12 @@ class Product
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(int $price): self
+    public function setPrice(float $price): self
     {
         $this->price = $price;
